+++ shared/public/Tiled2dMapRasterLayerInterface.h
@@ -37,9 +37,9 @@ public:
 
     virtual void removeCallbackHandler() = 0;
 
-    virtual void setAlpha(double alpha) = 0;
+    virtual void setAlpha(float alpha) = 0;
 
-    virtual double getAlpha() = 0;
+    virtual float getAlpha() = 0;
 
     virtual void setStyle(const ::RasterShaderStyle & style) = 0;
